% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_geo.R
\name{search_geo}
\alias{search_geo}
\title{Get Location Data from keyword}
\usage{
search_geo(search_val = NULL, return_geom = FALSE, address_details = FALSE)
}
\arguments{
\item{search_val}{Keyword(s) to be geocoded, e.g. column of postal codes}

\item{return_geom}{Default = \code{FALSE}. Whether to return the coordinate information}

\item{address_details}{Default = \code{FALSE}. Whether to return address information}
}
\description{
This function is a wrapper for the \href{https://www.onemap.gov.sg/apidocs/search/}{Search API}. It allows for geocoding of data such as postal codes or address information.
This is an internal function for the \code{geocode_onemap} function.
}
